move test_run_router() from test_em.py to test_run_router.py
